quote: extract simple cmdline arg quoting to utils
Extract the simple command-line argument quoting method from
AuthenticationBase to a utilitity class.
