+++ src/shared/Core/Authentication/AuthenticationBase.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -157,23 +156,6 @@ namespace GitCredentialManager.Authentication
             }
 
             return true;
-        }
-
-        public static string QuoteCmdArg(string str)
-        {
-            char[] needsQuoteChars = {'"', ' ', '\\', '\n', '\r', '\t'};
-            bool needsQuotes = str.Any(x => needsQuoteChars.Contains(x));
-
-            if (!needsQuotes)
-            {
-                return str;
-            }
-
-            // Replace all '\' characters with an escaped '\\', and all '"' with '\"'
-            string escapedStr = str.Replace("\\", "\\\\").Replace("\"", "\\\"");
-
-            // Bookend the escaped string with double-quotes '"'
-            return $"\"{escapedStr}\"";
         }
     }
 }
